Add podman macvlan network
Adds the host vlan interface "podnet".

Adds the podman network configuration file /etc/cni/net.d/podnetwork.conflist

With this I can launch podman containers with static IP on the podnet vlan.
Containers on this network will be directly visible on the network. And since
the host don't have an address in this network, it will reach the containers
via normal routing.

To launch containers on this network run:

   podman run -it --rm --network podnetwork --ip 172.20.2.109 \
	quay.io/libpod/banner /bin/sh

Possibly it's good to also specify a mac-address, so that arp caches will remain
valid when restarting containers?

   podman run -it --rm --network podnetwork --ip 172.20.2.109 \
	--mac-address 42:43:44:00:00:ae quay.io/libpod/banner /bin/sh